Horoscope-NFT-Project
In this course, you will learn about some super cool tools that are essential for interacting with smart contracts. You'll get to create your very first Web3 application and mint your own non-fungible token (NFT). You'll also learn how to interact with a deployed contract and connect Metamask to a website. Finally, you'll have the opportunity to ship your contract to a testnet and interact with it there as well. By the end of this course, you'll have a solid understanding of the tools and techniques needed to work with smart contracts and be well on your way to becoming a successful smart contract developer.